Sync license text
This patch updates metadata in `setup.py` and comments in other files to
clarify that the current license is AGPLv3, as is specified in the
current `LICENSE` file.

This patch also removes the "this agreement applies to this version and
all previous versions" text from the README and code comments, because
it is redundant; the previous code was already released under a _more
liberal_ license, so giving the option of AGPLv3 is not useful (and
perhaps confusing).