Remove fastText-related code
The motivation behind this change is to not have any untested code in
here that will just become out of sync with the rest of the repo and
become useless over time. If we really want a fastText example
somewhere, I'd suggest creating a `cleanlab-contrib` repo or similar
where we can keep untested code like this that may be useful to some
people.