Fix docs build in CI
This patch pins the Pandoc version to a more recent release (2.19.2).

As explained in d8cc2e50566e90dcfb9aff118a41610d4be070ee, nbconvert
translates Markdown -> reST -> HTML, and it uses Pandoc for this
Markdown -> reST conversion. Older versions of Pandoc don't do a
high-fidelity conversion; in particular, they don't have great support
for mixed Markdown/HTML content.

Consider the following Markdown, the core of how we use the details
disclosure element in several tutorial notebooks:

    <details><summary>Summary</summary>

    Body.

    </details>

Using Pandoc 2.19.2, this converts to the following reST:

    .. raw:: html

       <details>

    .. raw:: html

       <summary>

    Summary

    .. raw:: html

       </summary>

    Body.

    .. raw:: html

       </details>

This is a correct conversion. However, using Pandoc 2.5, which is the version
of Pandoc that was being used by CI prior to this patch (the version of Pandoc
available in the package repository for Ubuntu 20.04), it converts as follows:

    .. raw:: html

       <details>

    Summary

    Body.

    .. raw:: html

       </details>

The <summary> tag is lost in this conversion.

This patch switches to a more recent version of Pandoc to fix this issue.