refactor(multi-label): 🏷️ parametrize `NDArray`
A generic type variable, `T`, that represents the precision of a data type.

While the data types for probabilities and scores are identical, it conceptually makes sense to define separate type aliases for each.

`np` is only used in defining the floating data type, which is only used for type hints.