fix(datalab): 🐛 turn issues/issue_summary/info into properties in Datalab with setters
This simplifies the Datalab constructor.
The downside is that using @property and @setter decorators for so many attributes takes a lot of space.

One idea is to check if Datalab could become a dataclass in the future. Then, the properties can be turned into `field`s. IIRC, dataclass fields won't be frozen like properties (w/o setters).