flag near-duplicate issues based on items in near duplicate sets.
Previously, only the threshold set the issues boolean column, and the near duplicate sets were computed separately with the same heuristic.
The near-duplicate sets were updated to be symmetric, but the issues didn't update accordingly.
Now, only the threshold is used to compute the near-duplciate sets, so it's easy to flag all examples within a near-duplcite set to be considered a near-duplicate.

This is verified in the "# Check if all examples in the near_duplicate_set of a flagged example are also flagged" of the TestNearDuplicateSets.test_symmetric_and_flagged_consistency test.