Introduce helper classes for validating model outputs used as inputs to Datalab.find_issues
Instead of introducing new arguments that behave
identically to exisisting arguments, which are mutually
exlusive, we use a model-output hierarchy that allows
us to validate and process as intputs to Datalab based
on the provided task.

This is now only focused on unifying the `pred_probs´
and ´predictions´ arguments at the
user-facing level (Datalab, specifically IssueFinder-level),
without affecting the signatures of lower level classes
(like the IssueManagers).

This is strategically designed to accommodate a wider
range of model outputs in the future. This includes
extending these validation techniques to handle
complex outputs like 'features', 'pred_probs' in
datasets with higher dimensionality (e.g.,
segmentation) or datasets featuring non-normalized
rows (e.g., multi-label), thereby broadening the
applicability of Datalab in diverse machine learning
tasks.