Temporarily Revert to macOS 12 in CI due to Incompatibility with Python 3.8 and 3.9 (#1110)
Due to a migration in our CI environment from `macos-latest` to macOS 14 runner images, we've encountered a compatibility issue with Python 3.8 and 3.9. The new architecture of macOS 14 does not currently support these Python versions. While we expect future updates to resolve this, we will temporarily revert our CI to use macOS 12 to ensure continuity in our workflows.