Add _find_issues_inputs attributes in LabelIssueManager and OutlierIssueManager
This commit creates the `find_issues_inputs` attribute in the `LabelIssueManager` and `OutlierIssueManager` classes. The `find_issues_inputs` attribute is a dictionary that keeps track of the inputs used in the `find_issues` method. It will be included in the respective info dictionaries. It can be temporarily be used by DataMonitor to figure out if features or pred_probs were used to run the necessary checks or not.