Use different euclidean distance implementation in OutOfDistribution class for small number of examples for numerical stability
The decision to provide this for a smaller number of examples is due to performance concerns. This precise implementation will be up to 60x slower, but will avoid catastrophic cancellations. This is to handle edge-cases where all of the data points are REALLY close to each other (compared to their norms).